% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occu.r
\name{occu}
\alias{occu}
\title{Fit the MacKenzie et al. (2002) Occupancy Model}
\usage{
occu(formula, data, knownOcc = numeric(0), method = "BFGS",
  control = list())
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of detection and occupancy in that order.}

\item{data}{An \link[unmarked][UnmarkedFrameOccu] object}

\item{method}{Method used to solve the model. Either "stan", or a method used by \link[optimx][optimx].}

\item{control}{List of options to specify model fitting procedure. See Details.}

\item{kwnownOcc}{Vector of sites that are known to be occupied. These should be supplied as row numbers of the y matrix, eg, c(3,8) if sites 3 and 8 were known to be occupied a priori.}
}
\value{
\link[unmarkedFitOccu] object.
}
\description{
This function fits the single season occupancy model of MacKenzie et al (2002).
}
\author{
Jeffrey O. Hanson
}

